+++ com.unity.render-pipelines.core/Runtime/Lighting/ProbeVolume/ProbeVolumePerSceneData.cs
@@ -37,7 +37,8 @@ namespace UnityEngine.Rendering
 
         internal Dictionary<string, PerStateData> states = new();
 
-        string currentState = null;
+        bool assetLoaded = false;
+        string currentState = null, transitionState = null;
 
         /// <summary>
         /// OnAfterDeserialize implementation.
@@ -142,9 +143,22 @@ namespace UnityEngine.Rendering
         bool ResolveSharedCellData() => asset != null && asset.ResolveSharedCellData(cellSharedDataAsset, cellSupportDataAsset);
         bool ResolvePerStateCellData()
         {
-            if (currentState == null || !states.TryGetValue(currentState, out var data))
-                return false;
-            return asset.ResolvePerStateCellData(data.cellDataAsset, data.cellOptionalDataAsset);
+            int loadedCount = 0;
+            string state0 = transitionState != null ? transitionState : currentState;
+            string state1 = transitionState != null ? currentState : null;
+            if (state0 != null && states.TryGetValue(state0, out var data0))
+            {
+                if (asset.ResolvePerStateCellData(data0.cellDataAsset, data0.cellOptionalDataAsset, 0))
+                    loadedCount++;
+            }
+            if (state1 != null && states.TryGetValue(state1, out var data1))
+            {
+                if (asset.ResolvePerStateCellData(data1.cellDataAsset, data1.cellOptionalDataAsset, loadedCount))
+                    loadedCount++;
+            }
+            for (var i = 0; i < asset.cells.Length; ++i)
+                asset.cells[i].hasTwoStates = loadedCount == 2;
+            return loadedCount != 0;
         }
 
         internal void QueueAssetLoading()
@@ -154,6 +168,7 @@ namespace UnityEngine.Rendering
 
             var refVol = ProbeReferenceVolume.instance;
             refVol.AddPendingAssetLoading(asset);
+            assetLoaded = true;
 #if UNITY_EDITOR
             if (refVol.sceneData != null)
                 refVol.bakingProcessSettings = refVol.sceneData.GetBakeSettingsForScene(gameObject.scene);
@@ -164,38 +179,52 @@ namespace UnityEngine.Rendering
         {
             if (asset != null)
                 ProbeReferenceVolume.instance.AddPendingAssetRemoval(asset);
+            assetLoaded = false;
         }
 
         void OnEnable()
         {
             ProbeReferenceVolume.instance.RegisterPerSceneData(this);
 
-            ResolveSharedCellData();
             if (ProbeReferenceVolume.instance.sceneData != null)
-                SetBakingState(ProbeReferenceVolume.instance.bakingState);
+                Initialize();
             // otherwise baking state will be initialized in ProbeReferenceVolume.Initialize when sceneData is loaded
         }
 
         void OnDisable()
         {
-            OnDestroy();
+            QueueAssetRemoval();
+            currentState = transitionState = null;
             ProbeReferenceVolume.instance.UnregisterPerSceneData(this);
         }
 
-        void OnDestroy()
+        internal void Initialize()
         {
+            ResolveSharedCellData();
+
             QueueAssetRemoval();
-            currentState = null;
+            currentState = ProbeReferenceVolume.instance.sceneData.bakingState;
+            transitionState = null;
+            QueueAssetLoading();
         }
 
-        internal void SetBakingState(string state)
+        internal void UpdateBakingState(string state, string previousState)
         {
-            if (state == currentState)
+            if (asset == null)
                 return;
 
-            QueueAssetRemoval();
+            // if we just change state, don't need to queue anything
+            // Just load state cells from disk and wait for blending to stream updates to gpu
+            // When blending factor is < 0.5, streaming will upload cell from transition state
+            // After that, it will always upload cells from current state
+            // So gradually, all loaded cells, will be either blended towards new state, or replaced by streaming
+
             currentState = state;
-            QueueAssetLoading();
+            transitionState = previousState;
+            if (!assetLoaded)
+                QueueAssetLoading();
+            else if (!ResolvePerStateCellData())
+                QueueAssetRemoval();
         }
 
 #if UNITY_EDITOR