+++ com.unity.render-pipelines.high-definition/CHANGELOG.md
@@ -22,7 +22,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added time slicing support for realtime reflection probes.
 - Added denoising for the path tracer.
 - Added an initial version of under water rendering for the water system.
-- Added Asymmetric projection and Screen Coordinates Override frame settings. Adapted post effects to support Screen Coordinates Override. (Used, for example, to support Cluster Display.)
 - Added option to animate APV sample noise to smooth it out when TAA is enabled.
 - Added default DOTS compatible loading shader (MaterialLoading.shader)
 - Add #pragma editor_sync_compilation directive to MaterialError.shader
@@ -137,6 +136,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Fixed camera motion vector pass reading last frame depth texture
 - Fixed issue with shader graph custom velocity and VFX (case 1388149)
 - Fixed motion vector rendering with shader graph with planar primitive (case 1398313)
+- Fixed probe volume baking not generating any probes on mac.
 - Fix a few UX issues in APV.
 
 ## [14.0.0] - 2021-11-17